Changes to MessageHandler. MessageHandler now checks the permissions of the user before doing anything (Needs testing). SessionDatabase now has a 24hour expiry time which is checked when a user requests something. Tokens of user who login and don't logout properly when they re login a fresh token is issued over the top of the old token. This will still ensure that a user is able to login if something crashes on the control panel, but also ensure that only one user session can be valid (old token wont work).
Small change to Billboard class. scheduled field added, which is checked before a user can update the database.

Database messages have been changed to display only if a success or fail happens, and now doesn't show each connection and and what happened (I feel at this point the new error messages are more helpful). Lots more error messages have been added to the server, for example saying that the user cant login because of incorrect details, or that a user doesn't have the perms to do this action.